no more overflow (uint64_t)

+++ runtime.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 void compare_runtime(int capacity) {
-    int tablopairs[capacity];
+    uint64_t tablopairs[capacity];
     std::cout << "Pour un tableau de "<< capacity<<" elements, en C++..." << std::endl ;
     auto start_time_point=std::chrono::high_resolution_clock::now();
     for (int i=0;i<capacity;++i){
@@ -11,7 +11,7 @@ void compare_runtime(int capacity) {
     auto end_time_point=std::chrono::high_resolution_clock::now();
     std::cout << "Da tableau cree en " << std::chrono::duration_cast<std::chrono::microseconds>(end_time_point-start_time_point) << std::endl;
     auto sum_start_time_point= std::chrono::high_resolution_clock::now();
-    int sum=0;
+    uint64_t sum=0;
     for (int i=0;i<capacity;++i){
         sum+=tablopairs[i];
     }
@@ -19,6 +19,7 @@ void compare_runtime(int capacity) {
     std::cout << "Da somme faite en " << std::chrono::duration_cast<std::chrono::microseconds>(sum_end_time_point-sum_start_time_point) << std::endl;
     std::cout << "Anyway, da somme en question: " << sum << "\n";
 }
+
 int main() {
     compare_runtime(1000);
     compare_runtime(10000);
@@ -28,4 +29,4 @@ int main() {
 }
 
 //C'est bien plus rapide sur C++ que sur Python,
-//mais pour 1 000 000 d'éléments, il y a un problème d'overflow
+//et plus de problème d'overflow avec uint64_t
